+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Was popping an extra copy of the tutorial every time you pressed the button
 - Handle multipliers that are dice rolls, e.g. 1d4+2 Vampire Spawn (Note they must be in Foundry [[/r 1d4+2]] form  to be recognized)
 
+## v0.5.5
+- Add Display XP after Combat setting to display Encounter XP and XP per player token when you End Combat
+- Confirm compatible with Foundry 0.7.6
+
 ## v0.5.4
 - Spanish translation of dialogs and messages (thanks https://github.com/lozalojo)
 - Move asking about creating a Map Note to when you want to run the Encounter; otherwise put a warning in the dynamic section